feat(model): Moving parser to be an internal tool
